Refactor: Move semverCmp function to separate file
